# src/main/java/com/sprint/findex/apidocs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Findex API",
    "description": "가볍고 빠른 외부 API 연동 금융 분석 도구 API 문서",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://sprint-project-1196140422.ap-northeast-2.elb.amazonaws.com/sb/findex",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/auto-sync-configs/{id}": {
      "patch": {
        "tags": [
          "자동 연동 설정 API"
        ],
        "summary": "자동 연동 설정 수정",
        "description": "기존 자동 연동 설정을 수정합니다.",
        "operationId": "updateAutoSyncConfig",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "자동 연동 설정 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoSyncConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "수정할 자동 연동 설정을 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "자동 연동 설정 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AutoSyncConfigDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 설정 값 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auto-sync-configs": {
      "get": {
        "tags": [
          "자동 연동 설정 API"
        ],
        "summary": "자동 연동 설정 목록 조회",
        "description": "자동 연동 설정 목록을 조회합니다. 필터링, 정렬, 커서 기반 페이지네이션을 지원합니다.",
        "operationId": "getAutoSyncConfigList",
        "parameters": [
          {
            "name": "indexInfoId",
            "in": "query",
            "description": "지수 정보 ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "활성화 여부",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "idAfter",
            "in": "query",
            "description": "이전 페이지 마지막 요소 ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "커서 (다음 페이지 시작점)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "정렬 필드 (indexInfo.indexName, enabled)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "indexInfo.indexName"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "정렬 방향 (asc, desc)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "페이지 크기",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "자동 연동 설정 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageResponseAutoSyncConfigDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 필터 값 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "에러 응답",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "에러 발생 시간",
            "example": "2025-03-06T05:39:06.152068Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP 상태 코드",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "잘못된 요청입니다."
          },
          "details": {
            "type": "string",
            "description": "에러 상세 내용",
            "example": "부서 코드는 필수입니다."
          }
        }
      },
...
      "AutoSyncConfigUpdateRequest": {
        "type": "object",
        "description": "자동 연동 설정 수정 요청",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "활성화 여부",
            "example": true
          }
        },
        "required": [
          "enabled"
        ]
      },
      "AutoSyncConfigDto": {
        "type": "object",
        "description": "자동 연동 설정 DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "자동 연동 설정 ID",
            "example": 1
          },
          "indexInfoId": {
            "type": "integer",
            "format": "int64",
            "description": "지수 정보 ID",
            "example": 1
          },
          "indexClassification": {
            "type": "string",
            "description": "지수 분류명",
            "example": "KOSPI시리즈"
          },
          "indexName": {
            "type": "string",
            "description": "지수명",
            "example": "IT 서비스"
          },
          "enabled": {
            "type": "boolean",
            "description": "활성화 여부",
            "example": true
          }
        }
      },
...
      "CursorPageResponseAutoSyncConfigDto": {
        "type": "object",
        "description": "커서 기반 페이지 응답",
        "properties": {
          "content": {
            "type": "array",
            "description": "페이지 내용",
            "items": {
              "$ref": "#/components/schemas/AutoSyncConfigDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "다음 페이지 커서",
            "example": "eyJpZCI6MjB9"
          },
          "nextIdAfter": {
            "type": "integer",
            "format": "int64",
            "description": "마지막 요소의 ID",
            "example": "eyJpZCI6MjB9"
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "description": "페이지 크기",
            "example": 10
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "description": "총 요소 수",
            "example": 100
          },
          "hasNext": {
            "type": "boolean",
            "description": "다음 페이지 여부",
            "example": true
          }
        }
      }
    }
  }
}